Changed size of file input boxes
Updated size of label in `add_file_selection` to match the width of the
text box and button combined.
Reduced the size of the browse button and increased width of the text
box, still has the same total width.
Reworded the matrix utilities labels to fit within the UI width so they
aren't cut off.